plant post likes

+++ components/Comment/PlantComment.js
@@ -1,29 +1,55 @@
 import React, { useState, useEffect } from 'react';
 import { collection, addDoc, onSnapshot, doc, updateDoc, deleteDoc, getDoc } from 'firebase/firestore';
 import { db, auth } from '../../firebase';
+import PlantLikes from '../../components/Likes/PlantLikes'; // Ensure the path is correct
 import styles from './PlantComment.module.css';
-import Link from 'next/link';
-import PlantLikes from '../Likes/PlantLikes';
+import Modal from '../Modal/Modal';
+import Link from 'next/link'; // Import Link from next/link
 
-const PlantComment = ({ plantPostId }) => {
+const PlantComment = ({ postId, userId, plantId, plantPostId }) => {
     const [comments, setComments] = useState([]);
     const [newComment, setNewComment] = useState('');
     const [editingCommentId, setEditingCommentId] = useState(null);
     const [editingCommentContent, setEditingCommentContent] = useState('');
     const [loading, setLoading] = useState(false);
+    const [postOwnerId, setPostOwnerId] = useState('');
+    const [showLoginModal, setShowLoginModal] = useState(false);
     const [commentAuthors, setCommentAuthors] = useState({});
 
     useEffect(() => {
-        const unsubscribeComments = onSnapshot(collection(db, `plantPosts/${plantPostId}/comments`), (snapshot) => {
+        const fetchPostData = async () => {
+            if (!postId) {
+                console.error('Invalid postId:', postId);
+                return;
+            }
+            
+            try {
+                const postRef = doc(db, 'posts', postId);
+                const postDoc = await getDoc(postRef);
+
+                if (postDoc.exists()) {
+                    const postData = postDoc.data();
+                    setPostOwnerId(postData.userId);
+                } else {
+                    console.error('No post found with the provided postId:', postId);
+                }
+            } catch (error) {
+                console.error('Error fetching post data:', error);
+            }
+        };
+
+        const unsubscribeComments = onSnapshot(collection(db, `posts/${postId}/comments`), (snapshot) => {
             const commentsData = snapshot.docs.map(doc => ({ id: doc.id, ...doc.data() }));
             setComments(commentsData);
             fetchCommentAuthors(commentsData);
         });
 
+        fetchPostData();
+
         return () => {
             unsubscribeComments();
         };
-    }, [plantPostId]);
+    }, [postId]);
 
     const fetchCommentAuthors = async (commentsData) => {
         const authorPromises = commentsData.map(async (comment) => {
@@ -60,13 +86,13 @@ const PlantComment = ({ plantPostId }) => {
 
     const handleAddComment = async () => {
         if (!auth.currentUser) {
-            // Handle case where user is not logged in
+            setShowLoginModal(true);
             return;
         }
 
         setLoading(true);
         try {
-            const commentsCollectionRef = collection(db, `plantPosts/${plantPostId}/comments`);
+            const commentsCollectionRef = collection(db, `posts/${postId}/comments`);
             await addDoc(commentsCollectionRef, {
                 content: newComment,
                 createdAt: new Date(),
@@ -83,7 +109,7 @@ const PlantComment = ({ plantPostId }) => {
     const handleEditComment = async (commentId) => {
         setLoading(true);
         try {
-            const commentRef = doc(db, `plantPosts/${plantPostId}/comments`, commentId);
+            const commentRef = doc(db, `posts/${postId}/comments`, commentId);
             await updateDoc(commentRef, {
                 content: editingCommentContent,
             });
@@ -97,10 +123,19 @@ const PlantComment = ({ plantPostId }) => {
     };
 
     const handleDeleteComment = async (commentId, commentUserId) => {
+        if (!auth.currentUser) {
+            setShowLoginModal(true);
+            return;
+        }
+
         setLoading(true);
         try {
-            const commentRef = doc(db, `plantPosts/${plantPostId}/comments`, commentId);
-            await deleteDoc(commentRef);
+            const commentRef = doc(db, `posts/${postId}/comments`, commentId);
+            if (auth.currentUser.uid === postOwnerId || auth.currentUser.uid === commentUserId) {
+                await deleteDoc(commentRef);
+            } else {
+                console.error('You are not authorized to delete this comment.');
+            }
         } catch (error) {
             console.error('Error deleting comment:', error);
         } finally {
@@ -108,9 +143,22 @@ const PlantComment = ({ plantPostId }) => {
         }
     };
 
+    const handleCommentBoxClick = () => {
+        if (!auth.currentUser) {
+            setShowLoginModal(true);
+        }
+    };
+
+    const handleLoginSuccess = () => {
+        setShowLoginModal(false);
+    };
+
     return (
         <div className={styles.commentSection}>
-            <PlantLikes plantPostId={plantPostId} />
+            <Modal isOpen={showLoginModal} onClose={() => setShowLoginModal(false)} onSuccess={handleLoginSuccess} />
+            
+            <PlantLikes userId={userId} plantId={plantId} plantPostId={plantPostId} />
+
             <div className={styles.commentList}>
                 {comments.map(comment => (
                     <div key={comment.id} className={styles.comment}>
@@ -127,25 +175,6 @@ const PlantComment = ({ plantPostId }) => {
                                     <small className={styles.authorName}>{commentAuthors[comment.id].name || 'Unknown User'}</small>
                                 </Link>
                             </div>
-                        )}
-                        {(auth.currentUser?.uid === comment.userId) && (
-                            <button
-                                onClick={() => {
-                                    setEditingCommentId(comment.id);
-                                    setEditingCommentContent(comment.content);
-                                }}
-                                className={styles.commentButton}
-                            >
-                                Edit
-                            </button>
-                        )}
-                        {(auth.currentUser?.uid === comment.userId) && (
-                            <button
-                                onClick={() => handleDeleteComment(comment.id, comment.userId)}
-                                className={styles.commentButton}
-                            >
-                                Delete
-                            </button>
                         )}
                         {editingCommentId === comment.id ? (
                             <div>
@@ -168,6 +197,25 @@ const PlantComment = ({ plantPostId }) => {
                         ) : (
                             <span>{comment.content}</span>
                         )}
+                        {(auth.currentUser?.uid === comment.userId || auth.currentUser?.uid === postOwnerId) && auth.currentUser ? (
+                            <button
+                                onClick={() => handleDeleteComment(comment.id, comment.userId)}
+                                className={styles.commentButton}
+                            >
+                                Delete
+                            </button>
+                        ) : null}
+                        {auth.currentUser?.uid === comment.userId && (
+                            <button
+                                onClick={() => {
+                                    setEditingCommentId(comment.id);
+                                    setEditingCommentContent(comment.content);
+                                }}
+                                className={styles.commentButton}
+                            >
+                                Edit
+                            </button>
+                        )}
                     </div>
                 ))}
             </div>
@@ -179,18 +227,17 @@ const PlantComment = ({ plantPostId }) => {
                     onChange={(e) => setNewComment(e.target.value)}
                     placeholder="Add a comment..."
                     className={styles.commentInput}
-                    disabled={loading || !auth.currentUser}
+                    onClick={handleCommentBoxClick}
+                    disabled={loading}
                 />
                 <button
                     onClick={handleAddComment}
-                    disabled={loading || !newComment.trim() || !auth.currentUser}
+                    disabled={loading || !newComment.trim()}
                     className={styles.commentButton}
                 >
                     Post
                 </button>
             </div>
-
-            
         </div>
     );
 };
